+++ src/cpp/src/module_genai/modules/md_img_preprocess/vision_preprocess.cpp
@@ -3,16 +3,18 @@
 
 #include "vision_preprocess.hpp"
 
-#include "module_genai/modules/model/qwen3_vl/qwen3_vision_preprocess.hpp"
+#include "module_genai/modules/model/qwen3_5/vision_preprocess.hpp"
+#include "module_genai/modules/model/qwen3_vl/vision_preprocess.hpp"
 
 namespace ov::genai::module {
 
 VisionPreprocess::PTR VisionPreprocess::create(const std::filesystem::path& model_path, VLMModelType model_type) {
 	switch (model_type) {
 	case VLMModelType::QWEN3_VL:
+		return std::make_shared<Qwen3VisionPreprocess>(model_path, model_type);
 	case VLMModelType::QWEN3_5:
 	case VLMModelType::QWEN3_5_MOE:
-		return std::make_shared<Qwen3VisionPreprocess>(model_path, model_type);
+		return std::make_shared<Qwen3_5VisionPreprocess>(model_path, model_type);
 	default:
 		return nullptr;
 	}
