+++ src/cpp/src/module_genai/modules/model/qwen3_5/qwen3_5preprocessor.cpp
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "qwen3_5preprocessor.hpp"
+
 #include "openvino/core/except.hpp"
 #include "nlohmann/json.hpp"
 #include <fstream>
@@ -24,7 +25,7 @@ Qwen3_5Preprocessor::Qwen3_5Preprocessor(const std::filesystem::path &model_path
     : m_preprocess_config(Qwen3_5VisionPreprocessConfig::from_json_file(model_path / "preprocessor_config.json")),
       m_vision_config(Qwen3_5VisionConfig::from_json_file(model_path / "config.json")) {
     load_pos_embed_weight(model_path);
-    m_video_processor = std::make_shared<Qwen3VLVideoProcessor>(model_path);
+        // m_video_processor = std::make_shared<Qwen3_5VLVideoProcessor>(model_path);
 }
 
 Qwen3_5PreprocessorOutput Qwen3_5Preprocessor::preprocess(const ov::Tensor &images) {
@@ -178,7 +179,7 @@ Qwen3_5PreprocessorOutput Qwen3_5Preprocessor::preprocess(const ov::Tensor &imag
 }
 
 Qwen3_5PreprocessorOutput Qwen3_5Preprocessor::preprocess_video(const std::vector<ov::Tensor> &frames) {
-    m_video_processor->preprocess(frames);
+    // m_video_processor->preprocess(frames);
     return {};
 }
