+++ src/cpp/src/module_genai/modules/model/qwen3_5/qwen3_5preprocessor.hpp
@@ -4,14 +4,16 @@
 #pragma once
 #include <filesystem>
 #include <cstdint>
+#include <memory>
 #include <utility>
 #include <vector>
 #include <string>
 #include "openvino/runtime/tensor.hpp"
 #include "qwen3_5config.hpp"
-#include "module_genai/modules/model/qwen3_vl/qwen3_vl_video_processor.hpp"
 
 namespace ov::genai::module {
+
+class IVideoProcessor;
 
 struct Qwen3_5PreprocessorOutput {
     ov::Tensor pixel_values;
@@ -32,7 +34,6 @@ public:
     // Just keep an interface for batch preprocess, the implementation can be optimized later
     std::vector<Qwen3_5PreprocessorOutput> preprocess_videos(const std::vector<std::vector<ov::Tensor>> &batch_frames);
 private:
-    std::shared_ptr<Qwen3VLVideoProcessor> m_video_processor = nullptr;
     Qwen3_5VisionPreprocessConfig m_preprocess_config;
     Qwen3_5VisionConfig m_vision_config;
     ov::Tensor m_pos_embed_weight;
